ai-services/services/nsfw-detector/requirements.txt
flask==3.0.0
tensorflow==2.15.0
pillow==10.2.0
numpy==1.26.3
opencv-python-headless==4.9.0.80
gunicorn==21.2.0
prometheus-client==0.19.0
